requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
audioread==3.0.1
bcrypt==4.2.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
clang-format==17.0.6
click==8.1.8
colorlog==6.8.0
contourpy==1.3.1
cpplint==1.6.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepspeed==0.16.2
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
einops==0.8.0
email_validator==2.2.0
fastapi==0.115.6
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.16.1
flake8==3.8.2
flake8-bugbear==23.3.12
flake8-comprehensions==3.16.0
flake8-executable==2.1.3
flake8-pyi==20.5.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gradio==5.9.1
gradio_client==1.5.2
grpcio==1.68.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
identify==2.6.3
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-ollama==0.2.2
langchain-text-splitters==0.3.4
langid==1.1.6
langsmith==0.2.4
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.10.0
mccabe==0.6.1
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy==1.14.0
mypy-extensions==1.0.0
narwhals==1.19.0
networkx==3.4.2
ninja==1.11.1.3
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
ollama==0.4.4
openai==1.58.1
openai-whisper==20231117
orjson==3.10.12
packaging==23.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
pip==24.3.1
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
pre-commit==3.5.0
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycodestyle==2.6.0
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydantic-settings==2.7.0
pydub==0.25.1
pyflakes==2.2.0
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.8.4
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.12.0
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
TextGrid==1.6.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.28.0
websockets==11.0.3
Werkzeug==3.1.3
yarl==1.18.3